+++ x-pack/plugin/searchable-snapshots/src/main/java/org/elasticsearch/index/store/cache/CachedBlobContainerIndexInput.java
@@ -9,44 +9,26 @@ package org.elasticsearch.index.store.cache;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.apache.logging.log4j.message.ParameterizedMessage;
 import org.apache.lucene.codecs.CodecUtil;
 import org.apache.lucene.index.IndexFileNames;
-import org.apache.lucene.store.AlreadyClosedException;
 import org.apache.lucene.store.IOContext;
 import org.apache.lucene.store.IndexInput;
-import org.apache.lucene.util.BytesRef;
-import org.apache.lucene.util.BytesRefIterator;
-import org.elasticsearch.action.ActionListener;
-import org.elasticsearch.blobstore.cache.BlobStoreCacheService;
-import org.elasticsearch.blobstore.cache.CachedBlob;
-import org.elasticsearch.common.SuppressForbidden;
-import org.elasticsearch.common.bytes.BytesReference;
 import org.elasticsearch.common.collect.Tuple;
-import org.elasticsearch.common.io.Channels;
-import org.elasticsearch.common.lease.Releasable;
-import org.elasticsearch.common.unit.ByteSizeUnit;
 import org.elasticsearch.index.snapshots.blobstore.BlobStoreIndexShardSnapshot.FileInfo;
-import org.elasticsearch.index.store.BaseSearchableSnapshotIndexInput;
 import org.elasticsearch.index.store.IndexInputStats;
 import org.elasticsearch.index.store.SearchableSnapshotDirectory;
 import org.elasticsearch.xpack.searchablesnapshots.cache.ByteRange;
 
-import java.io.EOFException;
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.ByteBuffer;
-import java.nio.channels.FileChannel;
-import java.util.Locale;
 import java.util.concurrent.Future;
 import java.util.concurrent.atomic.AtomicLong;
-import java.util.concurrent.atomic.AtomicReference;
-import java.util.function.Consumer;
 import java.util.stream.IntStream;
 
 import static org.elasticsearch.xpack.searchablesnapshots.SearchableSnapshotsUtils.toIntBytes;
 
-public class CachedBlobContainerIndexInput extends BaseSearchableSnapshotIndexInput {
+public class CachedBlobContainerIndexInput extends MetadataCachingIndexInput {
 
     /**
      * Specific IOContext used for prewarming the cache. This context allows to write
@@ -56,22 +38,6 @@ public class CachedBlobContainerIndexInput extends BaseSearchableSnapshotIndexIn
     public static final IOContext CACHE_WARMING_CONTEXT = new IOContext();
 
     private static final Logger logger = LogManager.getLogger(CachedBlobContainerIndexInput.class);
-    private static final int COPY_BUFFER_SIZE = ByteSizeUnit.KB.toIntBytes(8);
-
-    private final CacheFileReference cacheFileReference;
-    private final int defaultRangeSize;
-    private final int recoveryRangeSize;
-
-    /**
-     * If > 0, represents a logical file within a compound (CFS) file or is a slice thereof represents the offset of the logical
-     * compound file within the physical CFS file
-     */
-    private final long compoundFileOffset;
-
-    // last read position is kept around in order to detect (non)contiguous reads for stats
-    private long lastReadPosition;
-    // last seek position is kept around in order to detect forward/backward seeks for stats
-    private long lastSeekPosition;
 
     public CachedBlobContainerIndexInput(
         String name,
@@ -97,7 +63,6 @@ public class CachedBlobContainerIndexInput extends BaseSearchableSnapshotIndexIn
             directory.getBlobCacheByteRange(name, fileInfo.length()),
             ByteRange.EMPTY
         );
-        assert getBufferSize() <= BlobStoreCacheService.DEFAULT_CACHED_BLOB_SIZE; // must be able to cache at least one buffer's worth
         stats.incrementOpenCount();
     }
 
@@ -111,286 +76,63 @@ public class CachedBlobContainerIndexInput extends BaseSearchableSnapshotIndexIn
         long compoundFileOffset,
         long length,
         CacheFileReference cacheFileReference,
-        int rangeSize,
+        int defaultRangeSize,
         int recoveryRangeSize,
         ByteRange headerBlobCacheByteRange,
         ByteRange footerBlobCacheByteRange
     ) {
-        super(logger, name, directory, fileInfo, context, stats, offset, length, headerBlobCacheByteRange, footerBlobCacheByteRange);
-        this.cacheFileReference = cacheFileReference;
-        this.lastReadPosition = this.offset;
-        this.lastSeekPosition = this.offset;
-        this.defaultRangeSize = rangeSize;
-        this.recoveryRangeSize = recoveryRangeSize;
-        this.compoundFileOffset = compoundFileOffset;
+        super(
+            logger,
+            name,
+            directory,
+            fileInfo,
+            context,
+            stats,
+            offset,
+            compoundFileOffset,
+            length,
+            cacheFileReference,
+            defaultRangeSize,
+            recoveryRangeSize,
+            headerBlobCacheByteRange,
+            footerBlobCacheByteRange
+        );
     }
 
     @Override
-    public void doClose() {
-        if (isClone == false) {
-            cacheFileReference.releaseOnClose();
-        }
-    }
-
-    private long getDefaultRangeSize() {
+    protected long getDefaultRangeSize() {
         return (context != CACHE_WARMING_CONTEXT)
             ? (directory.isRecoveryFinalized() ? defaultRangeSize : recoveryRangeSize)
             : fileInfo.partSize().getBytes();
     }
 
-    private ByteRange computeRange(long position) {
-        final long rangeSize = getDefaultRangeSize();
-        long start = (position / rangeSize) * rangeSize;
-        long end = Math.min(start + rangeSize, fileInfo.length());
-        return ByteRange.of(start, end);
-    }
-
     @Override
-    protected void doReadInternal(ByteBuffer b) throws IOException {
+    protected void readWithoutBlobCache(ByteBuffer b) throws Exception {
         ensureContext(ctx -> ctx != CACHE_WARMING_CONTEXT);
         final long position = getAbsolutePosition();
         final int length = b.remaining();
 
-        logger.trace("readInternal: read [{}-{}] ([{}] bytes) from [{}]", position, position + length, length, this);
-        try {
-            final CacheFile cacheFile = cacheFileReference.get();
+        final CacheFile cacheFile = cacheFileReference.get();
 
-            // Can we serve the read directly from disk? If so, do so and don't worry about anything else.
+        final ByteRange startRangeToWrite = computeRange(position);
+        final ByteRange endRangeToWrite = computeRange(position + length - 1);
+        assert startRangeToWrite.end() <= endRangeToWrite.end() : startRangeToWrite + " vs " + endRangeToWrite;
+        final ByteRange rangeToWrite = startRangeToWrite.minEnvelope(endRangeToWrite);
 
-            final Future<Integer> waitingForRead = cacheFile.readIfAvailableOrPending(ByteRange.of(position, position + length), chan -> {
-                final int read = readCacheFile(chan, position, b);
-                assert read == length : read + " vs " + length;
-                return read;
-            });
+        final ByteRange rangeToRead = ByteRange.of(position, position + length);
+        assert rangeToRead.isSubRangeOf(rangeToWrite) : rangeToRead + " vs " + rangeToWrite;
+        assert rangeToRead.length() == b.remaining() : b.remaining() + " vs " + rangeToRead;
 
-            if (waitingForRead != null) {
-                final Integer read = waitingForRead.get();
-                assert read == length;
-                readComplete(position, length);
-                return;
-            }
+        final Future<Integer> populateCacheFuture = cacheFile.populateAndRead(
+            rangeToWrite,
+            rangeToRead,
+            channel -> readCacheFile(channel, position, b),
+            this::writeCacheFile,
+            directory.cacheFetchAsyncExecutor()
+        );
 
-            // Requested data is not on disk, so try the cache index next.
-            final ByteRange indexCacheMiss; // null if not a miss
-
-            final ByteRange blobCacheByteRange = maybeReadFromBlobCache(position, length);
-            if (blobCacheByteRange.isEmpty() == false) {
-                final CachedBlob cachedBlob = directory.getCachedBlob(fileInfo.physicalName(), blobCacheByteRange);
-                assert cachedBlob == CachedBlob.CACHE_MISS || cachedBlob == CachedBlob.CACHE_NOT_READY || cachedBlob.from() <= position;
-                assert cachedBlob == CachedBlob.CACHE_MISS || cachedBlob == CachedBlob.CACHE_NOT_READY || length <= cachedBlob.length();
-
-                if (cachedBlob == CachedBlob.CACHE_MISS || cachedBlob == CachedBlob.CACHE_NOT_READY) {
-                    // We would have liked to find a cached entry but we did not find anything: the cache on the disk will be requested
-                    // so we compute the region of the file we would like to have the next time. The region is expressed as a tuple of
-                    // {start, end} where positions are relative to the whole file.
-                    indexCacheMiss = blobCacheByteRange;
-
-                    // We must fill in a cache miss even if CACHE_NOT_READY since the cache index is only created on the first put.
-                    // TODO TBD use a different trigger for creating the cache index and avoid a put in the CACHE_NOT_READY case.
-                } else {
-                    logger.trace(
-                        "reading [{}] bytes of file [{}] at position [{}] using cache index",
-                        length,
-                        fileInfo.physicalName(),
-                        position
-                    );
-                    stats.addIndexCacheBytesRead(cachedBlob.length());
-                    try {
-                        final int sliceOffset = toIntBytes(position - cachedBlob.from());
-                        final BytesRefIterator cachedBytesIterator = cachedBlob.bytes().slice(sliceOffset, length).iterator();
-                        BytesRef bytesRef;
-                        while ((bytesRef = cachedBytesIterator.next()) != null) {
-                            b.put(bytesRef.bytes, bytesRef.offset, bytesRef.length);
-                        }
-                        assert b.position() == length : "copied " + b.position() + " but expected " + length;
-
-                        final ByteRange cachedRange = ByteRange.of(cachedBlob.from(), cachedBlob.to());
-                        cacheFile.populateAndRead(
-                            cachedRange,
-                            cachedRange,
-                            channel -> cachedBlob.length(),
-                            (channel, from, to, progressUpdater) -> {
-                                final long startTimeNanos = stats.currentTimeNanos();
-                                final BytesRefIterator iterator = cachedBlob.bytes()
-                                    .slice(toIntBytes(from - cachedBlob.from()), toIntBytes(to - from))
-                                    .iterator();
-                                long writePosition = from;
-                                BytesRef current;
-                                while ((current = iterator.next()) != null) {
-                                    final ByteBuffer byteBuffer = ByteBuffer.wrap(current.bytes, current.offset, current.length);
-                                    while (byteBuffer.remaining() > 0) {
-                                        writePosition += positionalWrite(channel, writePosition, byteBuffer);
-                                        progressUpdater.accept(writePosition);
-                                    }
-                                }
-                                assert writePosition == to : writePosition + " vs " + to;
-                                final long endTimeNanos = stats.currentTimeNanos();
-                                stats.addCachedBytesWritten(to - from, endTimeNanos - startTimeNanos);
-                                logger.trace("copied bytes [{}-{}] of file [{}] from cache index to disk", from, to, fileInfo);
-                            },
-                            directory.cacheFetchAsyncExecutor()
-                        );
-                    } catch (Exception e) {
-                        logger.debug(
-                            new ParameterizedMessage(
-                                "failed to store bytes [{}-{}] of file [{}] obtained from index cache",
-                                cachedBlob.from(),
-                                cachedBlob.to(),
-                                fileInfo
-                            ),
-                            e
-                        );
-                        // oh well, no big deal, at least we can return them to the caller.
-                    }
-
-                    readComplete(position, length);
-
-                    return;
-                }
-            } else {
-                // requested range is not eligible for caching
-                indexCacheMiss = null;
-            }
-
-            // Requested data is also not in the cache index, so we must visit the blob store to satisfy both the target range and any
-            // miss in the cache index.
-
-            final ByteRange startRangeToWrite = computeRange(position);
-            final ByteRange endRangeToWrite = computeRange(position + length - 1);
-            assert startRangeToWrite.end() <= endRangeToWrite.end() : startRangeToWrite + " vs " + endRangeToWrite;
-
-            // only request what's needed to fill the cache to make sure the cache is always fully populated during recovery
-            final ByteRange rangeToWrite = indexCacheMiss != null && directory.isRecoveryFinalized() == false
-                ? indexCacheMiss
-                : startRangeToWrite.minEnvelope(endRangeToWrite);
-
-            final ByteRange rangeToRead = ByteRange.of(position, position + length);
-            assert rangeToRead.isSubRangeOf(rangeToWrite) : rangeToRead + " vs " + rangeToWrite;
-            assert rangeToRead.length() == b.remaining() : b.remaining() + " vs " + rangeToRead;
-
-            final Future<Integer> populateCacheFuture = cacheFile.populateAndRead(
-                rangeToWrite,
-                rangeToRead,
-                channel -> readCacheFile(channel, position, b),
-                this::writeCacheFile,
-                directory.cacheFetchAsyncExecutor()
-            );
-
-            if (indexCacheMiss != null) {
-                fillIndexCache(cacheFile, indexCacheMiss);
-                if (compoundFileOffset > 0L
-                    && indexCacheMiss.equals(headerBlobCacheByteRange)
-                    && footerBlobCacheByteRange.isEmpty() == false) {
-                    fillIndexCache(cacheFile, footerBlobCacheByteRange);
-                }
-            }
-
-            final int bytesRead = populateCacheFuture.get();
-            assert bytesRead == length : bytesRead + " vs " + length;
-        } catch (final Exception e) {
-            // may have partially filled the buffer before the exception was thrown, so try and get the remainder directly.
-            final int alreadyRead = length - b.remaining();
-            final int bytesRead = readDirectlyIfAlreadyClosed(position + alreadyRead, b, e);
-            assert alreadyRead + bytesRead == length : alreadyRead + " + " + bytesRead + " vs " + length;
-
-            // In principle we could handle an index cache miss here too, ensuring that the direct read was large enough, but this is
-            // already a rare case caused by an overfull/undersized cache.
-        }
-
-        readComplete(position, length);
-    }
-
-    private void fillIndexCache(CacheFile cacheFile, ByteRange indexCacheMiss) {
-        final Releasable onCacheFillComplete = stats.addIndexCacheFill();
-        final Future<Integer> readFuture = cacheFile.readIfAvailableOrPending(indexCacheMiss, channel -> {
-            final int indexCacheMissLength = toIntBytes(indexCacheMiss.length());
-
-            // We assume that we only cache small portions of blobs so that we do not need to:
-            // - use a BigArrays for allocation
-            // - use an intermediate copy buffer to read the file in sensibly-sized chunks
-            // - release the buffer once the indexing operation is complete
-
-            final ByteBuffer byteBuffer = ByteBuffer.allocate(indexCacheMissLength);
-            Channels.readFromFileChannelWithEofException(channel, indexCacheMiss.start(), byteBuffer);
-            // NB use Channels.readFromFileChannelWithEofException not readCacheFile() to avoid counting this in the stats
-            byteBuffer.flip();
-            final BytesReference content = BytesReference.fromByteBuffer(byteBuffer);
-            directory.putCachedBlob(fileInfo.physicalName(), indexCacheMiss.start(), content, new ActionListener<Void>() {
-                @Override
-                public void onResponse(Void response) {
-                    onCacheFillComplete.close();
-                }
-
-                @Override
-                public void onFailure(Exception e1) {
-                    onCacheFillComplete.close();
-                }
-            });
-            return indexCacheMissLength;
-        });
-
-        if (readFuture == null) {
-            // Normally doesn't happen, we're already obtaining a range covering all cache misses above, but theoretically
-            // possible in the case that the real populateAndRead call already failed to obtain this range of the file. In that
-            // case, simply move on.
-            onCacheFillComplete.close();
-        }
-    }
-
-    private void readComplete(long position, int length) {
-        stats.incrementBytesRead(lastReadPosition, position, length);
-        lastReadPosition = position + length;
-        lastSeekPosition = lastReadPosition;
-    }
-
-    private int readDirectlyIfAlreadyClosed(long position, ByteBuffer b, Exception e) throws IOException {
-        if (e instanceof AlreadyClosedException || (e.getCause() != null && e.getCause() instanceof AlreadyClosedException)) {
-            try {
-                // cache file was evicted during the range fetching, read bytes directly from blob container
-                final long length = b.remaining();
-                final byte[] copyBuffer = new byte[toIntBytes(Math.min(COPY_BUFFER_SIZE, length))];
-                logger.trace(
-                    () -> new ParameterizedMessage(
-                        "direct reading of range [{}-{}] for cache file [{}]",
-                        position,
-                        position + length,
-                        cacheFileReference
-                    )
-                );
-
-                int bytesCopied = 0;
-                final long startTimeNanos = stats.currentTimeNanos();
-                try (InputStream input = openInputStreamFromBlobStore(position, length)) {
-                    long remaining = length;
-                    while (remaining > 0) {
-                        final int len = (remaining < copyBuffer.length) ? (int) remaining : copyBuffer.length;
-                        int bytesRead = input.read(copyBuffer, 0, len);
-                        if (bytesRead == -1) {
-                            throw new EOFException(
-                                String.format(
-                                    Locale.ROOT,
-                                    "unexpected EOF reading [%d-%d] ([%d] bytes remaining) from %s",
-                                    position,
-                                    position + length,
-                                    remaining,
-                                    cacheFileReference
-                                )
-                            );
-                        }
-                        b.put(copyBuffer, 0, bytesRead);
-                        bytesCopied += bytesRead;
-                        remaining -= bytesRead;
-                        assert remaining == b.remaining() : remaining + " vs " + b.remaining();
-                    }
-                    final long endTimeNanos = stats.currentTimeNanos();
-                    stats.addDirectBytesRead(bytesCopied, endTimeNanos - startTimeNanos);
-                }
-                return bytesCopied;
-            } catch (Exception inner) {
-                e.addSuppressed(inner);
-            }
-        }
-        throw new IOException("failed to read data from cache", e);
+        final int bytesRead = populateCacheFuture.get();
+        assert bytesRead == length : bytesRead + " vs " + length;
     }
 
     /**
@@ -474,44 +216,6 @@ public class CachedBlobContainerIndexInput extends BaseSearchableSnapshotIndexIn
         }
     }
 
-    @SuppressForbidden(reason = "Use positional writes on purpose")
-    private static int positionalWrite(FileChannel fc, long start, ByteBuffer byteBuffer) throws IOException {
-        assert assertCurrentThreadMayWriteCacheFile();
-        return fc.write(byteBuffer, start);
-    }
-
-    /**
-     * Perform a single {@code read()} from {@code inputStream} into {@code copyBuffer}, handling an EOF by throwing an {@link EOFException}
-     * rather than returning {@code -1}. Returns the number of bytes read, which is always positive.
-     *
-     * Most of its arguments are there simply to make the message of the {@link EOFException} more informative.
-     */
-    private static int readSafe(
-        InputStream inputStream,
-        byte[] copyBuffer,
-        long rangeStart,
-        long rangeEnd,
-        long remaining,
-        CacheFileReference cacheFileReference
-    ) throws IOException {
-        final int len = (remaining < copyBuffer.length) ? toIntBytes(remaining) : copyBuffer.length;
-        final int bytesRead = inputStream.read(copyBuffer, 0, len);
-        if (bytesRead == -1) {
-            throw new EOFException(
-                String.format(
-                    Locale.ROOT,
-                    "unexpected EOF reading [%d-%d] ([%d] bytes remaining) from %s",
-                    rangeStart,
-                    rangeEnd,
-                    remaining,
-                    cacheFileReference
-                )
-            );
-        }
-        assert bytesRead > 0 : bytesRead;
-        return bytesRead;
-    }
-
     /**
      * Asserts that the range of bytes to warm in cache is aligned with {@link #fileInfo}'s part size.
      */
@@ -528,60 +232,6 @@ public class CachedBlobContainerIndexInput extends BaseSearchableSnapshotIndexIn
                 + "] is not aligned with part end or with file length";
         }
         return true;
-    }
-
-    private int readCacheFile(final FileChannel fc, final long position, final ByteBuffer buffer) throws IOException {
-        assert assertFileChannelOpen(fc);
-        final int bytesRead = Channels.readFromFileChannel(fc, position, buffer);
-        if (bytesRead == -1) {
-            throw new EOFException(
-                String.format(
-                    Locale.ROOT,
-                    "unexpected EOF reading [%d-%d] from %s",
-                    position,
-                    position + buffer.remaining(),
-                    cacheFileReference
-                )
-            );
-        }
-        stats.addCachedBytesRead(bytesRead);
-        return bytesRead;
-    }
-
-    private void writeCacheFile(final FileChannel fc, final long start, final long end, final Consumer<Long> progressUpdater)
-        throws IOException {
-        assert assertFileChannelOpen(fc);
-        assert assertCurrentThreadMayWriteCacheFile();
-        final long length = end - start;
-        final byte[] copyBuffer = new byte[toIntBytes(Math.min(COPY_BUFFER_SIZE, length))];
-        logger.trace(() -> new ParameterizedMessage("writing range [{}-{}] to cache file [{}]", start, end, cacheFileReference));
-
-        long bytesCopied = 0L;
-        long remaining = end - start;
-        final long startTimeNanos = stats.currentTimeNanos();
-        try (InputStream input = openInputStreamFromBlobStore(start, length)) {
-            while (remaining > 0L) {
-                final int bytesRead = readSafe(input, copyBuffer, start, end, remaining, cacheFileReference);
-                positionalWrite(fc, start + bytesCopied, ByteBuffer.wrap(copyBuffer, 0, bytesRead));
-                bytesCopied += bytesRead;
-                remaining -= bytesRead;
-                progressUpdater.accept(start + bytesCopied);
-            }
-            final long endTimeNanos = stats.currentTimeNanos();
-            stats.addCachedBytesWritten(bytesCopied, endTimeNanos - startTimeNanos);
-        }
-    }
-
-    @Override
-    protected void seekInternal(long pos) throws IOException {
-        if (pos > length()) {
-            throw new EOFException("Reading past end of file [position=" + pos + ", length=" + length() + "] for " + toString());
-        } else if (pos < 0L) {
-            throw new IOException("Seeking to negative position [" + pos + "] for " + toString());
-        }
-        final long position = pos + this.offset;
-        stats.incrementSeeks(lastSeekPosition, position);
-        lastSeekPosition = position;
     }
 
     @Override
@@ -619,7 +269,7 @@ public class CachedBlobContainerIndexInput extends BaseSearchableSnapshotIndexIn
         if (sliceCompoundFile) {
             sliceCompoundFileOffset = this.offset + sliceOffset;
             sliceHeaderByteRange = directory.getBlobCacheByteRange(sliceName, sliceLength).shift(sliceCompoundFileOffset);
-            if (sliceHeaderByteRange.length() < sliceLength) {
+            if (sliceHeaderByteRange.isEmpty() == false && sliceHeaderByteRange.length() < sliceLength) {
                 sliceFooterByteRange = ByteRange.of(sliceLength - CodecUtil.footerLength(), sliceLength).shift(sliceCompoundFileOffset);
             } else {
                 sliceFooterByteRange = ByteRange.EMPTY;
@@ -665,83 +315,5 @@ public class CachedBlobContainerIndexInput extends BaseSearchableSnapshotIndexIn
                 )
                 : null)
             + ']';
-    }
-
-    private static class CacheFileReference implements CacheFile.EvictionListener {
-
-        private final long fileLength;
-        private final CacheKey cacheKey;
-        private final SearchableSnapshotDirectory directory;
-        private final AtomicReference<CacheFile> cacheFile = new AtomicReference<>(); // null if evicted or not yet acquired
-
-        private CacheFileReference(SearchableSnapshotDirectory directory, String fileName, long fileLength) {
-            this.cacheKey = directory.createCacheKey(fileName);
-            this.fileLength = fileLength;
-            this.directory = directory;
-        }
-
-        CacheFile get() throws Exception {
-            CacheFile currentCacheFile = cacheFile.get();
-            if (currentCacheFile != null) {
-                return currentCacheFile;
-            }
-
-            final CacheFile newCacheFile = directory.getCacheFile(cacheKey, fileLength);
-            synchronized (this) {
-                currentCacheFile = cacheFile.get();
-                if (currentCacheFile != null) {
-                    return currentCacheFile;
-                }
-                newCacheFile.acquire(this);
-                final CacheFile previousCacheFile = cacheFile.getAndSet(newCacheFile);
-                assert previousCacheFile == null;
-                return newCacheFile;
-            }
-        }
-
-        @Override
-        public void onEviction(final CacheFile evictedCacheFile) {
-            synchronized (this) {
-                if (cacheFile.compareAndSet(evictedCacheFile, null)) {
-                    evictedCacheFile.release(this);
-                }
-            }
-        }
-
-        void releaseOnClose() {
-            synchronized (this) {
-                final CacheFile currentCacheFile = cacheFile.getAndSet(null);
-                if (currentCacheFile != null) {
-                    currentCacheFile.release(this);
-                }
-            }
-        }
-
-        @Override
-        public String toString() {
-            return "CacheFileReference{"
-                + "cacheKey='"
-                + cacheKey
-                + '\''
-                + ", fileLength="
-                + fileLength
-                + ", acquired="
-                + (cacheFile.get() != null)
-                + '}';
-        }
-    }
-
-    private static boolean assertFileChannelOpen(FileChannel fileChannel) {
-        assert fileChannel != null;
-        assert fileChannel.isOpen();
-        return true;
-    }
-
-    private static boolean assertCurrentThreadMayWriteCacheFile() {
-        final String threadName = Thread.currentThread().getName();
-        assert isCacheFetchAsyncThread(threadName) : "expected the current thread ["
-            + threadName
-            + "] to belong to the cache fetch async thread pool";
-        return true;
     }
 }
