+++ x-pack/plugin/searchable-snapshots/src/test/java/org/elasticsearch/index/store/cache/TestUtils.java
@@ -26,6 +26,7 @@ import org.elasticsearch.common.bytes.BytesReference;
 import org.elasticsearch.common.io.PathUtilsForTesting;
 import org.elasticsearch.common.io.Streams;
 import org.elasticsearch.common.settings.Settings;
+import org.elasticsearch.common.unit.ByteSizeValue;
 import org.elasticsearch.index.store.IndexInputStats;
 import org.elasticsearch.xpack.searchablesnapshots.cache.ByteRange;
 
@@ -325,6 +326,11 @@ public final class TestUtils {
         @Override
         protected void getAsync(String repository, String name, String path, long offset, ActionListener<CachedBlob> listener) {
             listener.onResponse(CachedBlob.CACHE_NOT_READY);
+        }
+
+        @Override
+        public ByteRange computeBlobCacheByteRange(String fileName, long fileLength, ByteSizeValue maxMetadataLength) {
+            return ByteRange.EMPTY;
         }
 
         @Override
